Ship FakeFileProvider even with dist dev releases
Refs psalm/psalm.dev#60
